requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
argilla==1.12.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backoff==2.2.1
bangla==0.0.2
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
boltons==23.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.4
clip-anytorch==2.5.2
cmake==3.26.4
commonmark==0.9.1
contourpy==1.1.0
coqpit==0.0.17
cryptography==41.0.1
cycler==0.11.0
Cython==0.29.30
dataclasses-json==0.5.9
dateparser==1.1.8
decorator==5.1.1
Deprecated==1.2.14
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
# Editable Git install with no remote (EmoSupport==0.0.1)
encodec==0.1.1
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
g2pkk==0.1.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.56.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
inflect==5.6.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.1
jsonlines==1.2.0
jsonmerge==1.9.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
k-diffusion==0.0.15
kiwisolver==1.4.4
kornia==0.6.12
langchain==0.0.228
langchainplus-sdk==0.0.20
lazy_loader==0.3
librosa==0.10.0.post2
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
mdurl==0.1.2
mecab-python3==1.0.6
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.57.0
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
openai==0.27.8
openai-whisper==20230314
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pathtools==0.1.2
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
pooch==1.6.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pynndescent==0.5.10
pypandoc==1.11
pyparsing==3.0.9
pypinyin==0.49.0
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rfc3986==1.5.0
rich==13.0.1
rpds-py==0.8.10
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.27.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.5
SQLAlchemy==2.0.18
streamlit==1.24.1
streamlit-audiorecorder==0.0.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.1.0
tifffile==2023.7.4
tiktoken==0.3.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
trainer==0.0.27
trampoline==0.1.2
transformers==4.30.2
triton==2.0.0
TTS==0.15.6
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
umap-learn==0.5.1
unidic-lite==1.0.8
unstructured==0.8.0
urllib3==1.26.16
validators==0.20.0
wandb==0.15.5
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.0
-e .